+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.0.14 (2023-09-05)
+## 0.0.14 (2023-09-06)
 
 
 ### Bug Fixes
@@ -6,10 +6,8 @@
 * build issue ([89b7bab](https://gitlab.com/paras205/boilerplate-react/commit/89b7bab7ed5f7fe63870961d67610dc0bfacf3c6))
 * colour, design and api response changes ([72990a9](https://gitlab.com/paras205/boilerplate-react/commit/72990a9e3a85297001f6462d54cfa1c5722eb6d3))
 * hotfix: route manage ([9c0394c](https://gitlab.com/paras205/boilerplate-react/commit/9c0394ce422983cccca2b7e439c6f699faf9c135))
-* member update issue ([3dd2304](https://gitlab.com/paras205/boilerplate-react/commit/3dd2304dba5eaf291a0aceb717b13c1054c9e1dd))
 * refactor code ([7e26a3b](https://gitlab.com/paras205/boilerplate-react/commit/7e26a3bc30eb42c9bb385f334b448c7aaccd916e))
 * remove comment ([af0713c](https://gitlab.com/paras205/boilerplate-react/commit/af0713cfc99343cee95b131be67ba43855644b6d))
-* service update ([d2af844](https://gitlab.com/paras205/boilerplate-react/commit/d2af844e467eadb7f27a536a048a5f6a0749f5ff))
 
 
 ### Features
@@ -17,17 +15,26 @@
 * add additional folder ([6a3e80e](https://gitlab.com/paras205/boilerplate-react/commit/6a3e80e0283bb533435461c064b30935cd953ddf))
 * add and member list completed ([934b4ea](https://gitlab.com/paras205/boilerplate-react/commit/934b4eab6da534d90b24e3b7fe2844ec5de668b4))
 * added password component, refactored input component ([5676063](https://gitlab.com/paras205/boilerplate-react/commit/5676063ecf247dd38bd7be2ca29c19d3b9c6d76a))
+* added react-scroll ([13fede6](https://gitlab.com/paras205/boilerplate-react/commit/13fede605a505aba9d8e6fd138442513e206ec12))
 * added set password template ([75f9471](https://gitlab.com/paras205/boilerplate-react/commit/75f9471f735d8a6cf734924dcc21e0f56326750a))
 * adjusted column width, added edit and add modal in member tier ([827e90e](https://gitlab.com/paras205/boilerplate-react/commit/827e90ebb34d3c5ee73046ad70f05feac8f87be0))
 * admin loagin integartion ([6a58586](https://gitlab.com/paras205/boilerplate-react/commit/6a5858668898a187d75d9633e45aef0ec1867ac1))
 * api integration for add member tier ([591e472](https://gitlab.com/paras205/boilerplate-react/commit/591e4724e8b1fbd33f620dc73ed81df0beed9334))
+* created changes on profile ([87224a2](https://gitlab.com/paras205/boilerplate-react/commit/87224a2175f8773709c0bf174bf33c1d1f5a10c1))
+* created changes on profile header sticky ([d255f28](https://gitlab.com/paras205/boilerplate-react/commit/d255f285f625a0c90fabd871e093319899091d78))
+* created changes on profile header sticky ([6856523](https://gitlab.com/paras205/boilerplate-react/commit/6856523acbc56a000de611b3efe084d28baa3d4b))
+* created changes on profile header sticky ([64e7ef0](https://gitlab.com/paras205/boilerplate-react/commit/64e7ef04fbcf1d43102c3e32c4e83ca6a69f5f66))
+* created changes on profile page sections ([9686c6c](https://gitlab.com/paras205/boilerplate-react/commit/9686c6c04225187bc4ae3b86ecf5da742cee7d36))
+* created profile  design changes ([1a58f5d](https://gitlab.com/paras205/boilerplate-react/commit/1a58f5d57998cc0d393985b210a086ce8e8210d5))
 * custom input ([d323c8f](https://gitlab.com/paras205/boilerplate-react/commit/d323c8f84f00f51287e6c564fa41478402ce5ae9))
+* disabled select if already selected earlier ([17ec76c](https://gitlab.com/paras205/boilerplate-react/commit/17ec76ce3e20f68927377bf1318cee4017490877))
 * forgot password page design and api implementation ([83fdd11](https://gitlab.com/paras205/boilerplate-react/commit/83fdd1147575e825dfdff059120828e07843d96e))
 * header navbar ([db616a8](https://gitlab.com/paras205/boilerplate-react/commit/db616a8aa8ca6b88e5900b684437b8ebf9b3f345))
 * login component ([f753932](https://gitlab.com/paras205/boilerplate-react/commit/f753932002845eaab1196b776d992ee5ecac0547))
 * login component ([967e159](https://gitlab.com/paras205/boilerplate-react/commit/967e159a4fe973d5e6302f108f2bd08525fcc615))
 * login component ([9da8c3e](https://gitlab.com/paras205/boilerplate-react/commit/9da8c3ed3142a97a439caf5080ed56106b689458))
 * logot session clear from font side ([29ac586](https://gitlab.com/paras205/boilerplate-react/commit/29ac5861e27e9b8d2e15e114cc47f0b540ea00ee))
+* password validation ([a14954c](https://gitlab.com/paras205/boilerplate-react/commit/a14954c5935baa2dcd783af303ed3b8ea450aab0))
 * refetched point for manual and added phone validation ([1d51211](https://gitlab.com/paras205/boilerplate-react/commit/1d51211fd510cf093aab1d4739037ff4e30ab572))
 * resolved modal issue, integrated api for update property ([56fc7b3](https://gitlab.com/paras205/boilerplate-react/commit/56fc7b3dce0062608340685fa86dcee7ab0d45da))